plugin: Move eventQueueSet to its own file
